GfxBackend: Commands are now 16-bit.

+++ src/wondergfx/wg_backendlogger.h
@@ -67,7 +67,7 @@ namespace wg
 		void	setRects(const RectSPX* pBeg, const RectSPX* pEnd) override;
 		void	setColors(const HiColor* pBeg, const HiColor* pEnd) override;
 		void	setTransforms(const Transform * pBeg, const Transform * pEnd) override;
-		void	processCommands(const int32_t* pBeg, const int32_t* pEnd) override;
+		void	processCommands(const uint16_t* pBeg, const uint16_t* pEnd) override;
 
 		//.____ Misc _________________________________________________________
